Add files via upload
This code creates a simple RNN(recurrent neural network) using Keras to generate handwritten-like text. It starts with a small dataset of example sentences, maps characters to numbers, and builds an LSTM model. After training, it allows users to input a starting phrase and generates text based on that input.